Refactor entity IDs from int to Long for consistency across models

+++ src/main/java/com/schoolhealth/schoolmedical/entity/Disease.java
@@ -17,7 +17,7 @@ public class Disease {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "disease_id")
-    private int diseaseId;
+    private Long diseaseId;
 
     @Column(name = "name", nullable = false, length = 255)
     private String name;
