+++ src/main/java/com/schoolhealth/schoolmedical/entity/VaccinationHistory.java
@@ -18,7 +18,7 @@ public class VaccinationHistory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "history_id")
-    private int historyId;
+    private Long historyId;
 
     @Column(name = "vaccine_id")
     private int vaccineId;
