+++ src/main/java/com/schoolhealth/schoolmedical/repository/HealthCheckConsentRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HealthCheckConsentRepo extends JpaRepository<HealthCheckConsentForm, Integer> {
+public interface HealthCheckConsentRepo extends JpaRepository<HealthCheckConsentForm, Long> {
 }
