+++ src/main/java/com/schoolhealth/schoolmedical/repository/HealthCheckHistoryRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HealthCheckHistoryRepo extends JpaRepository<HealthCheckHistory, Integer> {
+public interface HealthCheckHistoryRepo extends JpaRepository<HealthCheckHistory, Long> {
 }
